Time: 0 ms (100.00%) | Memory: 7.2 MB (7.24%) - LeetSync

+++ 342-power-of-four/power-of-four.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPowerOfFour(int n) {
         if (n<=0) return 0;
-        double exp=log2(n);
+        float exp=log2(n);
         if (exp!=(int)exp) return 0;
         return ((int)exp%2==0);
     }
